User commit

+++ src/main/java/com/example/demo/repository/ValidationRepository.java
@@ -3,7 +3,7 @@ import org.springframework.stereotype.Repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import com.example.demo.entity.ValidationEntity;
 @Repository
-public interface ValidationRepository extends JpaRepository<ValidationEntity,Integer>{
+public interface ValidationRepository extends JpaRepository<ValidationEntity,Long>{
     
 
 }
